Intensity Integrated to classifier/ New data downloaded
I've finished integerating the intensity code to the classifier.
I've experimented with our training set, and want to move on to the new data.
I've changed the data directory root, and the classifier is now throwing format related errors. Change the root to original to avoid this error.
The data is too big to be uploaded to github.