Now you can test out the cause feedback yourself!
The feedback cause is exported in .pkl format.
All you have to do is change loadFile to True on the last line of pipeline.py
run(loadFile=True, causeFilename="((cm,-0.4)=0.84)causeSunWed",outputDivider=900)

for changing settings for the accuracy tests, edit line 59

outputDividor will divid the total number of tweets and output only a few of the tweets you are loading.